requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# face-recognition==1.3.0  # Requer CMake no Windows - instale separadamente
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
SQLAlchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2

